Pre-Release 1.4.9
= Updated Webinterface Design (Not release yet).
= Updated TimeZone footer on LoggingMessages.
= Updated Webinterface to send the Link Privat.
= Fixed some typos.
= Fixed error when Ree6 doesn't have the Perms to delete a message.
= Some messages are deleting themselves now.

+++ src/main/java/de/presti/ree6/commands/Command.java
@@ -3,6 +3,7 @@ package de.presti.ree6.commands;
 import de.presti.ree6.utils.Logger;
 
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.MessageChannel;
@@ -86,7 +87,7 @@ public abstract class Command {
     }
 
     public static void deleteMessage(Message message) {
-        if(message != null) {
+        if(message != null && message.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_MANAGE)) {
             try {
                 message.delete().queue();
             } catch (Exception ex) {
